Complete Refactoring, make Number Crunching Code standalone.
A big step towards making a Stella Library to be used by other Riecoin Software, reducing Code Duplication, and also for other useful applications that may make use of Prime Constellations.

Version Number is going to follow the same scheme as Riecoin Core, so it will be 24.10, but let it be 0.94 for now.

Signed-off-by: Pttn <28868425+Pttn@users.noreply.github.com>